updated mixed-mode binaries and source
