+++ mixed-mode/EmptyClassDll/main.cs
@@ -1,0 +1,1 @@
+class MixedMode {}
