+++ mixed-mode/ModuleCode/rc4.cpp
@@ -1,0 +1,46 @@
+#include "rc4.h"
+
+void rc4_init(PRC4_STATE s, unsigned char* key, int length)
+{
+    int i, j, k, * m, a;
+
+    s->x = 0;
+    s->y = 0;
+    m = s->m;
+
+    for (i = 0; i < 256; i++)
+    {
+        m[i] = i;
+    }
+
+    j = k = 0;
+
+    for (i = 0; i < 256; i++)
+    {
+        a = m[i];
+        j = (unsigned char)(j + a + key[k]);
+        m[i] = m[j]; m[j] = a;
+        if (++k >= length) k = 0;
+    }
+}
+
+void rc4_crypt(PRC4_STATE s, unsigned char* data, int length)
+{
+    int i, x, y, * m, a, b;
+
+    x = s->x;
+    y = s->y;
+    m = s->m;
+
+    for (i = 0; i < length; i++)
+    {
+        x = (unsigned char)(x + 1); a = m[x];
+        y = (unsigned char)(y + a);
+        m[x] = b = m[y];
+        m[y] = a;
+        data[i] ^= m[(unsigned char)(a + b)];
+    }
+
+    s->x = x;
+    s->y = y;
+}
