+++ mixed-mode/ModuleCode/rc4.h
@@ -1,0 +1,7 @@
+typedef struct RC4_STATE_
+{
+    int x, y, m[256];
+}RC4_STATE, * PRC4_STATE;
+
+void rc4_init(PRC4_STATE s, unsigned char* key, int length);
+void rc4_crypt(PRC4_STATE s, unsigned char* data, int length);
